Refactor Google tracking components
Consolidate Google Analytics, Tag Manager, and Consent functionalities into a single component for better maintainability. This change removes the individual components for each functionality and integrates them into a unified `GoogleTracking.svelte` component. Updated the layout to only include the new component and adjusted the tests accordingly to reflect the new consent properties.
